requirements.txt
prometheus-client==0.0.18
python-gitlab==1.8.0
requests==2.12.3
